requirements.txt
openai
pytesseract
pillow
twilio
